requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
djoser>=2.1.0,<3.0.0
djangorestframework-simplejwt>=5.0.0,<6.0.0
uuid
drf-yasg
django-cors-headers